Q20 GRU upgrade v2: clean redo with all files consistent
LUT (gen_lut_tables.m → ulunas_lut.h/.c):
- 4096pt Q20 sigmoid/tanh LUTs (power-of-2 range: 2^24, 2^23)
- GTCRN-style bitwise index/frac (pos>>SHIFT, pos&(RANGE-1))
- Inline sigmoid_q20_to_q20 / tanh_q20_to_q20

GRU (ulunas_fp.c/h):
- gru_step_fp_q20: int32_t Q20 hidden state, Q20 gates, Q20 mixing
  Q20→Q15 output: (hn+16)>>5 for backward compat
- gru_sequence_fp_q20, bigru_sequence_fp_q20
- sat_s20() helper

State (ulunas_fp.h):
- tfa_cache_e0-e4/d0-d4: int16_t→int32_t Q20
- inter_cache_0/1: int16_t→int32_t Q20

Callers updated:
- ctfa_ta_fp: int32_t ta_h_cache + gru_step_fp_q20
- ctfa_fa_fp: bigru_sequence_fp_q20
- ctfa wrapper: int32_t tc
- encoder_xconv_module: int32_t tfa_cache
- gdprnn_module: bigru_sequence_fp_q20 + gru_sequence_fp_q20

Golden/diag regenerated from same MATLAB session.

+++ x2000_deploy_v2/ulunas_modules.c
@@ -137,7 +137,7 @@ void skip_add(const int32_t *a, const int32_t *b, int N, int32_t *y) {
 /* ======================================================================== */
 
 void encoder_xconv_module(
-    const int32_t *x, int32_t *conv_cache, int16_t *tfa_cache, int32_t *y);
+    const int32_t *x, int32_t *conv_cache, int32_t *tfa_cache, int32_t *y);
 
 /* ======================================================================== */
 /*  ENCODER LAYER 2: XDWS0  [24,33]→[24,33]                                 */
@@ -594,7 +594,7 @@ void gdprnn_module(
 
 /* Encoder Layer 0 — moved to top of file for dependencies */
 void encoder_xconv_module(
-    const int32_t *x, int32_t *conv_cache, int16_t *tfa_cache, int32_t *y)
+    const int32_t *x, int32_t *conv_cache, int32_t *tfa_cache, int32_t *y)
 {
     int32_t xc[3*129];
     for(int i=0;i<E0_CACHE_FRAMES*E0_IN_W;i++)xc[i]=conv_cache[i];
